+++ JavaSpring-2018-11-HomeWork-14/src/main/java/ru/otus/homework/models/jpa/Genre.java
@@ -4,9 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import ru.otus.homework.models.DataSet;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigInteger;
 
 @Data
 @NoArgsConstructor
@@ -20,7 +22,7 @@ public class Genre implements Serializable, DataSet
     @Id
     @Column(name = "genre_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private BigInteger id;
 
     @Column(unique = true)
     private String value;
